@@ -38,11 +38,11 @@ task clean(type: Delete) {
 }
 
 ext {
-    compileSdkVersion = 26
-    buildToolsVersion = '25.0.0'
-    targetSdkVersion = 26
-    minSdkVersion = 23
+    compileSdkVersion = 28
+    buildToolsVersion = '28.0.3'
+    targetSdkVersion = 28
+    minSdkVersion = 22
     versionCode = 1
     versionName = "1.0"
-    supportVersion = '27.1.1'
+    supportVersion = '28.0.0'
 }
