@@ -1,15 +1,17 @@
 package com.yc.mugua.presenter;
 
-import android.os.Handler;
-
 import com.blankj.utilcode.util.StringUtils;
+import com.lzy.okgo.model.Response;
 import com.yc.mugua.R;
+import com.yc.mugua.bean.BaseResponseBean;
 import com.yc.mugua.bean.DataBean;
+import com.yc.mugua.callback.Code;
+import com.yc.mugua.controller.CloudApi;
 import com.yc.mugua.impl.FeedbackContract;
-import com.yc.mugua.impl.ForgetContract;
 
-import java.util.ArrayList;
-import java.util.List;
+import io.reactivex.Observer;
+import io.reactivex.android.schedulers.AndroidSchedulers;
+import io.reactivex.disposables.Disposable;
 
 /**
  * Created by Android Studio.
@@ -21,26 +23,69 @@ public class FeedbackPresenter extends FeedbackContract.Presenter {
 
     @Override
     public void onRequest() {
-        new Handler().postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                List<DataBean> list = new ArrayList<>();
-                for (int i = 0;i<15;i++){
-                    list.add(new DataBean());
-                }
-                mView.setData(list);
-                mView.hideLoading();
-            }
-        }, 500);
+        CloudApi.mineReasons()
+                .doOnSubscribe(disposable -> {})
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribe(new Observer<Response<BaseResponseBean<DataBean>>>() {
+                    @Override
+                    public void onSubscribe(Disposable d) {
+                        mView.addDisposable(d);
+                    }
+
+                    @Override
+                    public void onNext(Response<BaseResponseBean<DataBean>> baseResponseBeanResponse) {
+                        if (baseResponseBeanResponse.body().code == Code.CODE_SUCCESS){
+                            DataBean data = baseResponseBeanResponse.body().data;
+                            mView.setData(data.getReasons());
+                        }
+                    }
+
+                    @Override
+                    public void onError(Throwable e) {
+                        mView.onError(e);
+                    }
+
+                    @Override
+                    public void onComplete() {
+                        mView.hideLoading();
+                    }
+                });
     }
 
     @Override
-    public void onFeed(String toString) {
-        if (StringUtils.isEmpty(toString)){
+    public void onFeed(String text, String id) {
+        if (StringUtils.isEmpty(text) || StringUtils.isEmpty(id)) {
             showToast(act.getString(R.string.error_));
             return;
         }
-        mView.finish();
+        CloudApi.mineOpinion(text, id)
+                .doOnSubscribe(disposable -> {
+                })
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribe(new Observer<Response<BaseResponseBean>>() {
+                    @Override
+                    public void onSubscribe(Disposable d) {
+                        mView.addDisposable(d);
+                    }
+
+                    @Override
+                    public void onNext(Response<BaseResponseBean> baseResponseBeanResponse) {
+                        if (baseResponseBeanResponse.body().code == Code.CODE_SUCCESS) {
+                            mView.finish();
+                        }
+                        showToast(baseResponseBeanResponse.body().message);
+                    }
+
+                    @Override
+                    public void onError(Throwable e) {
+                        mView.onError(e);
+                    }
+
+                    @Override
+                    public void onComplete() {
+                        mView.hideLoading();
+                    }
+                });
     }
 
 }
