@@ -1,14 +1,17 @@
 package com.yc.mugua.presenter;
 
-import android.os.Handler;
-
+import com.lzy.okgo.model.Response;
 import com.yc.mugua.base.BaseFragment;
+import com.yc.mugua.bean.BaseResponseBean;
 import com.yc.mugua.bean.DataBean;
+import com.yc.mugua.callback.Code;
+import com.yc.mugua.controller.CloudApi;
 import com.yc.mugua.controller.UIHelper;
 import com.yc.mugua.impl.SearchContract;
 
-import java.util.ArrayList;
-import java.util.List;
+import io.reactivex.Observer;
+import io.reactivex.android.schedulers.AndroidSchedulers;
+import io.reactivex.disposables.Disposable;
 
 /**
  * Created by Android Studio.
@@ -19,32 +22,34 @@ import java.util.List;
 public class SearchPresenter extends SearchContract.Presenter{
     @Override
     public void onHot() {
-        new Handler().postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                List<DataBean> list = new ArrayList<>();
-                for (int i = 0;i<5;i++){
-                    list.add(new DataBean());
-                }
-                mView.setHot(list);
-                mView.hideLoading();
-            }
-        }, 500);
-    }
+        CloudApi.homeKeywords()
+                .doOnSubscribe(disposable -> {})
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribe(new Observer<Response<BaseResponseBean<DataBean>>>() {
+                    @Override
+                    public void onSubscribe(Disposable d) {
+                        mView.addDisposable(d);
+                    }
 
-    @Override
-    public void onRecommend() {
-        new Handler().postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                List<DataBean> list = new ArrayList<>();
-                for (int i = 0;i<5;i++){
-                    list.add(new DataBean());
-                }
-                mView.setRecommend(list);
-                mView.hideLoading();
-            }
-        }, 500);
+                    @Override
+                    public void onNext(Response<BaseResponseBean<DataBean>> baseResponseBeanResponse) {
+                        if (baseResponseBeanResponse.body().code == Code.CODE_SUCCESS){
+                            DataBean data = baseResponseBeanResponse.body().data;
+                            mView.setHot(data.getHot());
+                            mView.setRecommend(data.getRecommend());
+                        }
+                    }
+
+                    @Override
+                    public void onError(Throwable e) {
+                        mView.onError(e);
+                    }
+
+                    @Override
+                    public void onComplete() {
+                        mView.hideLoading();
+                    }
+                });
     }
 
     @Override
@@ -54,16 +59,32 @@ public class SearchPresenter extends SearchContract.Presenter{
 
     @Override
     public void onRequest(int pagetNumber, String text) {
-        new Handler().postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                List<DataBean> list = new ArrayList<>();
-                for (int i = 0;i<15;i++){
-                    list.add(new DataBean());
-                }
-                mView.setData(list);
-                mView.hideLoading();
-            }
-        }, 500);
+        CloudApi.findList(pagetNumber, text, null, null, null)
+                .doOnSubscribe(disposable -> {})
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribe(new Observer<Response<BaseResponseBean<DataBean>>>() {
+                    @Override
+                    public void onSubscribe(Disposable d) {
+                        mView.addDisposable(d);
+                    }
+
+                    @Override
+                    public void onNext(Response<BaseResponseBean<DataBean>> baseResponseBeanResponse) {
+                        if (baseResponseBeanResponse.body().code == Code.CODE_SUCCESS){
+                            DataBean data = baseResponseBeanResponse.body().data;
+                            mView.setData(data.getVideos());
+                        }
+                    }
+
+                    @Override
+                    public void onError(Throwable e) {
+                        mView.onError(e);
+                    }
+
+                    @Override
+                    public void onComplete() {
+                        mView.hideLoading();
+                    }
+                });
     }
 }
