@@ -6,7 +6,6 @@ import android.os.Handler;
 import android.support.v7.widget.AppCompatEditText;
 import android.text.Editable;
 import android.text.TextWatcher;
-import android.view.KeyEvent;
 import android.view.View;
 import android.view.inputmethod.EditorInfo;
 import android.view.inputmethod.InputMethodManager;
@@ -17,6 +16,7 @@ import com.flyco.roundview.RoundViewDelegate;
 import com.lcodecore.tkrefreshlayout.RefreshListenerAdapter;
 import com.lcodecore.tkrefreshlayout.TwinklingRefreshLayout;
 import com.yc.mugua.R;
+import com.yc.mugua.adapter.BillboardAdapter;
 import com.yc.mugua.base.BaseFragment;
 import com.yc.mugua.bean.DataBean;
 import com.yc.mugua.databinding.FSearchBinding;
@@ -25,6 +25,7 @@ import com.yc.mugua.presenter.SearchPresenter;
 import com.zhy.view.flowlayout.FlowLayout;
 import com.zhy.view.flowlayout.TagAdapter;
 
+import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -35,6 +36,11 @@ import java.util.List;
  *  搜索
  */
 public class SearchFrg extends BaseFragment<SearchPresenter, FSearchBinding> implements SearchContract.View, View.OnClickListener {
+
+    private List<DataBean> listBean = new ArrayList<>();
+    private String searchText;
+    private BillboardAdapter adapter;
+    private AppCompatEditText etSearch;
 
     @Override
     public void initPresenter() {
@@ -53,6 +59,7 @@ public class SearchFrg extends BaseFragment<SearchPresenter, FSearchBinding> imp
 
     @Override
     protected void initView(View view) {
+        etSearch = view.findViewById(R.id.et_search);
         AppCompatEditText etSearch = view.findViewById(R.id.et_search);
         etSearch.setFocusable(true);
         etSearch.setFocusableInTouchMode(true);
@@ -61,34 +68,37 @@ public class SearchFrg extends BaseFragment<SearchPresenter, FSearchBinding> imp
         mB.fyClose.setOnClickListener(this);
 
         showLoadDataing();
-        mB.refreshLayout.startRefresh();
+//        mB.refreshLayout.startRefresh();
         mPresenter.onHot();
-        mPresenter.onRecommend();
+        if (adapter == null){
+            adapter = new BillboardAdapter(act, listBean);
+        }
+        setRecyclerViewGridType(mB.recyclerView, 2, 60, 20, R.color.blue_15163d);
+        mB.recyclerView.setAdapter(adapter);
+        mB.refreshLayout.setEnableLoadmore(false);
         setRefreshLayout(mB.refreshLayout, new RefreshListenerAdapter() {
             @Override
             public void onRefresh(TwinklingRefreshLayout refreshLayout) {
-
+                mPresenter.onRequest(pagerNumber = 1, searchText);
             }
 
             @Override
             public void onLoadMore(TwinklingRefreshLayout refreshLayout) {
                 super.onLoadMore(refreshLayout);
+                mPresenter.onRequest(pagerNumber += 1, searchText);
             }
         });
-        etSearch.setOnEditorActionListener(new TextView.OnEditorActionListener() {
-            @Override
-            public boolean onEditorAction(TextView v, int actionId, KeyEvent event) {
-                //判断是否是“完成”键
-                if(actionId == EditorInfo.IME_ACTION_SEARCH){
-                    //隐藏软键盘
-                    InputMethodManager imm = (InputMethodManager) v.getContext().getSystemService(Context.INPUT_METHOD_SERVICE);
-                    if (imm.isActive()) {
-                        imm.hideSoftInputFromWindow(v.getApplicationWindowToken(), 0);
-                    }
-                    return true;
+        etSearch.setOnEditorActionListener((v, actionId, event) -> {
+            //判断是否是“完成”键
+            if(actionId == EditorInfo.IME_ACTION_SEARCH){
+                //隐藏软键盘
+                InputMethodManager imm = (InputMethodManager) v.getContext().getSystemService(Context.INPUT_METHOD_SERVICE);
+                if (imm.isActive()) {
+                    imm.hideSoftInputFromWindow(v.getApplicationWindowToken(), 0);
                 }
-                return false;
+                return true;
             }
+            return false;
         });
         etSearch.addTextChangedListener(new TextWatcher() {
             @Override
@@ -103,18 +113,16 @@ public class SearchFrg extends BaseFragment<SearchPresenter, FSearchBinding> imp
 
             @Override
             public void afterTextChanged(final Editable editable) {
-                new Handler().postDelayed(new Runnable() {
-                    @Override
-                    public void run() {
-                        if (editable.length() == 0){
-                            mB.tvSearch.setText("搜索");
-                        }else {
-                            mB.tvSearch.setText("取消");
-                        }
+                new Handler().postDelayed(() -> {
+                    if (editable.length() == 0){
+                        mB.tvSearch.setText("搜索");
+                    }else {
+                        mB.tvSearch.setText("取消");
                     }
                 }, 300);
             }
         });
+
     }
 
     @Override
@@ -123,10 +131,10 @@ public class SearchFrg extends BaseFragment<SearchPresenter, FSearchBinding> imp
         mB.flHot.removeAllViews();
         mB.flHot.setAdapter(new TagAdapter<DataBean>(list){
             @Override
-            public View getView(FlowLayout parent, int position, DataBean dataBean) {
+            public View getView(FlowLayout parent, int position, DataBean bean) {
                 View view = View.inflate(act, R.layout.i_search_label, null);
                 TextView tvText = view.findViewById(R.id.tv_text);
-                tvText.setText(position + "全部");
+                tvText.setText(bean.getName());
                 return view;
             }
 
@@ -135,10 +143,11 @@ public class SearchFrg extends BaseFragment<SearchPresenter, FSearchBinding> imp
                 super.onSelected(position, view);
                 RoundTextView tvText = view.findViewById(R.id.tv_text);
                 RoundViewDelegate delegate = tvText.getDelegate();
-                delegate.setBackgroundColor(act.getColor(R.color.red_F72A61));
+                delegate.setBackgroundColor(act.getResources().getColor(R.color.red_F72A61));
                 DataBean bean = list.get(position);
                 bean.setSelect(true);
-                mPresenter.onSearch(SearchFrg.this);
+                searchText = bean.getName();
+                mB.refreshLayout.startRefresh();
             }
 
             @Override
@@ -158,10 +167,10 @@ public class SearchFrg extends BaseFragment<SearchPresenter, FSearchBinding> imp
         mB.flRecommend.removeAllViews();
         mB.flRecommend.setAdapter(new TagAdapter<DataBean>(list){
             @Override
-            public View getView(FlowLayout parent, int position, DataBean dataBean) {
+            public View getView(FlowLayout parent, int position, DataBean bean) {
                 View view = View.inflate(act, R.layout.i_search_label, null);
                 TextView tvText = view.findViewById(R.id.tv_text);
-                tvText.setText(position + "全部");
+                tvText.setText(bean.getName());
                 return view;
             }
 
@@ -173,7 +182,8 @@ public class SearchFrg extends BaseFragment<SearchPresenter, FSearchBinding> imp
                 delegate.setBackgroundColor(act.getColor(R.color.red_F72A61));
                 DataBean bean = list.get(position);
                 bean.setSelect(true);
-                mPresenter.onSearch(SearchFrg.this);
+                searchText = bean.getName();
+                mB.refreshLayout.startRefresh();
             }
 
             @Override
@@ -190,19 +200,36 @@ public class SearchFrg extends BaseFragment<SearchPresenter, FSearchBinding> imp
 
     @Override
     public void setRefreshLayoutMode(int totalRow) {
+        super.setRefreshLayoutMode(listBean.size(), totalRow, mB.refreshLayout);
+    }
 
+    @Override
+    public void hideLoading() {
+        super.hideLoading();
+        super.setRefreshLayout(pagerNumber, mB.refreshLayout);
     }
 
     @Override
     public void setData(Object data) {
-
+        mB.recyclerView.setVisibility(View.VISIBLE);
+        mB.layout.setVisibility(View.GONE);
+        List<DataBean> list = (List<DataBean>) data;
+        if (pagerNumber == 1) {
+            listBean.clear();
+            mB.refreshLayout.finishRefreshing();
+        } else {
+            mB.refreshLayout.finishLoadmore();
+        }
+        listBean.addAll(list);
+        adapter.notifyDataSetChanged();
     }
 
     @Override
     public void onClick(View view) {
         switch (view.getId()){
             case R.id.tv_search:
-
+                searchText = etSearch.getText().toString().trim();
+                mB.refreshLayout.startRefresh();
                 break;
             case R.id.fy_close:
                 pop();
