@@ -16,7 +16,6 @@ import com.zhy.view.flowlayout.FlowLayout;
 import com.zhy.view.flowlayout.TagAdapter;
 import com.zhy.view.flowlayout.TagFlowLayout;
 
-import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -35,18 +34,14 @@ public class SearchAdapter extends BaseRecyclerviewAdapter<DataBean> {
     protected void onBindViewHolde(RecyclerView.ViewHolder holder, int position) {
         ViewHolder viewHolder = (ViewHolder) holder;
         DataBean bean = listBean.get(position);
-
-        final List<DataBean> list = new ArrayList<>();
-        list.add(new DataBean());
-        list.add(new DataBean());
-        list.add(new DataBean());
+        final List<DataBean> list = bean.getCategory();
         viewHolder.fl_search.removeAllViews();
         viewHolder.fl_search.setAdapter(new TagAdapter<DataBean>(list){
             @Override
             public View getView(FlowLayout parent, int position, DataBean dataBean) {
                 View view = View.inflate(act, R.layout.i_search_label, null);
                 TextView tvText = view.findViewById(R.id.tv_text);
-                tvText.setText(position + "全部");
+                tvText.setText(dataBean.getName());
                 return view;
             }
 
@@ -58,6 +53,9 @@ public class SearchAdapter extends BaseRecyclerviewAdapter<DataBean> {
                 DataBean bean = list.get(position);
                 delegate.setBackgroundColor(act.getColor(R.color.red_F72A61));
                 bean.setSelect(true);
+                if (listener != null){
+                    listener.onClick(position, null, bean.getId());
+                }
             }
 
             @Override
@@ -68,8 +66,19 @@ public class SearchAdapter extends BaseRecyclerviewAdapter<DataBean> {
                 DataBean bean = list.get(position);
                 delegate.setBackgroundColor(0);
                 bean.setSelect(false);
+                if (listener != null){
+                    listener.onClick(position, null, null);
+                }
             }
         });
+    }
+
+    private OnClickListener listener;
+    public void setListener(OnClickListener listener){
+        this.listener = listener;
+    }
+    public interface OnClickListener{
+        void onClick(int position, String keyword, String categoryId);
     }
 
     @Override
