@@ -19,12 +19,10 @@ import android.app.ProgressDialog;
 import android.content.Context;
 import android.databinding.DataBindingUtil;
 import android.databinding.ViewDataBinding;
-import android.os.Build;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Message;
 import android.support.annotation.Nullable;
-import android.support.annotation.RequiresApi;
 import android.support.v4.content.ContextCompat;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.DefaultItemAnimator;
@@ -52,12 +50,12 @@ import com.yc.mugua.event.PayInEvent;
 import com.yc.mugua.utils.TUtil;
 import com.yc.mugua.utils.pay.PayResult;
 import com.yc.mugua.weight.GridDividerItemDecoration;
-import com.yc.mugua.weight.LoadingLayout;
 
 import org.greenrobot.eventbus.EventBus;
 
 import java.util.Map;
 
+import ezy.ui.layout.LoadingLayout;
 import io.reactivex.disposables.CompositeDisposable;
 import io.reactivex.disposables.Disposable;
 import me.yokeyword.fragmentation_swipeback.SwipeBackFragment;
@@ -397,12 +395,12 @@ public abstract class BaseFragment<P extends BasePresenter, VB extends ViewDataB
     protected void setTitleTransparent(String title, boolean back){
         setTitle(title);
         view.findViewById(R.id.top_view).setVisibility(View.GONE);
-        view.findViewById(R.id.title_bar).setBackgroundColor(act.getColor(R.color.transparent));
+        view.findViewById(R.id.title_bar).setBackgroundColor(act.getResources().getColor(R.color.transparent));
         Toolbar toolbar = view.findViewById(R.id.toolbar);
         if (back){
             toolbar.setNavigationIcon(null);
         }
-        toolbar.setBackgroundColor(act.getColor(R.color.transparent));
+        toolbar.setBackgroundColor(act.getResources().getColor(R.color.transparent));
     }
 
     private void title(String title, String rightText, int img, boolean isBack) {
@@ -444,30 +442,34 @@ public abstract class BaseFragment<P extends BasePresenter, VB extends ViewDataB
     }
 
 
-    @RequiresApi(api = Build.VERSION_CODES.M)
     protected void setRecyclerViewType(RecyclerView recyclerView){
         recyclerView.setLayoutManager(new LinearLayoutManager(act));
         setRecyclerView(recyclerView, R.color.blue_15163d);
     }
 
-    @RequiresApi(api = Build.VERSION_CODES.M)
     protected void setRecyclerViewType(RecyclerView recyclerView, int baColor){
         recyclerView.setLayoutManager(new LinearLayoutManager(act));
         setRecyclerView(recyclerView, baColor);
     }
 
-    @RequiresApi(api = Build.VERSION_CODES.M)
     protected void setRecyclerViewGridType(RecyclerView recyclerView, int spanCount, int width, int height, int color){
         recyclerView.setLayoutManager(new GridLayoutManager(act, spanCount));
         recyclerView.addItemDecoration(new GridDividerItemDecoration(width, height, ContextCompat.getColor(act,color)));
         setRecyclerView(recyclerView, color);
     }
 
-    @RequiresApi(api = Build.VERSION_CODES.M)
     protected void setRecyclerViewGridType(RecyclerView recyclerView, int spanCount, int width, int height){
         recyclerView.setLayoutManager(new GridLayoutManager(act, spanCount));
         recyclerView.addItemDecoration(new GridDividerItemDecoration(width, height, ContextCompat.getColor(act,R.color.white)));
         setRecyclerView(recyclerView, R.color.white);
+    }
+
+    public static void setMargins (View v, int l, int t, int r, int b) {
+        if (v.getLayoutParams() instanceof ViewGroup.MarginLayoutParams) {
+            ViewGroup.MarginLayoutParams p = (ViewGroup.MarginLayoutParams) v.getLayoutParams();
+            p.setMargins(l, t, r, b);
+            v.requestLayout();
+        }
     }
 
     private void setRecyclerView(RecyclerView recyclerView, int baColor){
