@@ -157,7 +157,7 @@ public class JsonConvert<T> implements Converter<T> {
                     throw new IllegalStateException("用户收取信息已过期");
                 } else {
                     //直接将服务端的错误信息抛出，onError中可以获取
-                    throw new IllegalStateException("错误代码：" + code + "，错误信息：" + lzyResponse.desc);
+                    throw new IllegalStateException("错误代码：" + code + "，错误信息：" + lzyResponse.message);
                 }
             }
         }
