@@ -14,11 +14,10 @@ package com.yc.mugua.callback;/*
  * limitations under the License.
  */
 
-import com.blankj.utilcode.util.Utils;
+import com.blankj.utilcode.util.ToastUtils;
 import com.google.gson.stream.JsonReader;
 import com.lzy.okgo.callback.AbsCallback;
 import com.yc.mugua.bean.BaseResponseBean;
-import com.yc.mugua.utils.cache.ShareSessionIdCache;
 
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
@@ -51,21 +50,16 @@ public abstract class NewsCallback<T> extends AbsCallback<T> {
         Type rawType = ((ParameterizedType) type).getRawType();
         if (rawType == BaseResponseBean.class) {
             BaseResponseBean gankResponse = Convert.fromJson(jsonReader, type);
-            if (gankResponse.code == 1) {
+            if (gankResponse.code == Code.CODE_SUCCESS) {
                 response.close();
                 //noinspection unchecked
                 return (T) gankResponse;
-            }else if (gankResponse.code == 0){
-//                ToastUtils.showShort(gankResponse.desc);
+            }else if (gankResponse.code == Code.CODE_FAILED){
+                ToastUtils.showShort(gankResponse.message);
                 response.close();
-                //noinspection unchecked
-                return (T) gankResponse;
+                throw new IllegalStateException(gankResponse.message);
             } else if (gankResponse.code == 2){
                 response.close();
-//                UIHelper.startSplashAct();
-//                ActivityUtils.startActivity(LoginAct.class);
-                ShareSessionIdCache.getInstance(Utils.getApp()).remove();
-//                ActivityUtils.finishAllActivities();
                 throw new IllegalStateException("用户在第三方登录");
             }else {
                 response.close();
