@@ -1,8 +1,17 @@
 package com.yc.mugua.presenter;
 
 import com.blankj.utilcode.util.StringUtils;
+import com.lzy.okgo.model.Response;
 import com.yc.mugua.R;
+import com.yc.mugua.bean.BaseResponseBean;
+import com.yc.mugua.bean.DataBean;
+import com.yc.mugua.callback.Code;
+import com.yc.mugua.controller.CloudApi;
 import com.yc.mugua.impl.IncomeContract;
+
+import io.reactivex.Observer;
+import io.reactivex.android.schedulers.AndroidSchedulers;
+import io.reactivex.disposables.Disposable;
 
 /**
  * Created by Android Studio.
@@ -13,11 +22,73 @@ import com.yc.mugua.impl.IncomeContract;
 public class IncomePresenter extends IncomeContract.Presenter{
 
     @Override
-    public void onSubmit(String name, String moble, String balance) {
+    public void onSubmit(String name, String moble, String balance, double withd) {
+        if (Double.valueOf(balance) > withd){
+            showToast("可提现余额不足");
+            return;
+        }
         if (StringUtils.isEmpty(name) || StringUtils.isEmpty(moble) || StringUtils.isEmpty(balance)){
             showToast(act.getString(R.string.error_));
             return;
         }
+        CloudApi.shareSubmit(moble, name, balance)
+                .doOnSubscribe(disposable -> {mView.showLoading();})
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribe(new Observer<Response<BaseResponseBean>>() {
+                    @Override
+                    public void onSubscribe(Disposable d) {
+                        mView.addDisposable(d);
+                    }
+
+                    @Override
+                    public void onNext(Response<BaseResponseBean> baseResponseBeanResponse) {
+                        if (baseResponseBeanResponse.body().code == Code.CODE_SUCCESS){
+                            mView.setShare();
+                        }
+                        showToast(baseResponseBeanResponse.body().message);
+                    }
+
+                    @Override
+                    public void onError(Throwable e) {
+                        mView.onError(e);
+                    }
+
+                    @Override
+                    public void onComplete() {
+                            mView.hideLoading();
+                    }
+                });
+    }
+
+    @Override
+    public void onInfo() {
+        CloudApi.shareInfo()
+                .doOnSubscribe(disposable -> {mView.showLoading();})
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribe(new Observer<Response<BaseResponseBean<DataBean>>>() {
+                    @Override
+                    public void onSubscribe(Disposable d) {
+                        mView.addDisposable(d);
+                    }
+
+                    @Override
+                    public void onNext(Response<BaseResponseBean<DataBean>> baseResponseBeanResponse) {
+                        if (baseResponseBeanResponse.body().code == Code.CODE_SUCCESS){
+                            DataBean data = baseResponseBeanResponse.body().data;
+                            mView.setData(data);
+                        }
+                    }
+
+                    @Override
+                    public void onError(Throwable e) {
+                        mView.onError(e);
+                    }
+
+                    @Override
+                    public void onComplete() {
+                        mView.hideLoading();
+                    }
+                });
     }
 
 }
