@@ -9,7 +9,7 @@ import android.os.Environment;
 public class Constants {
 
    public static final int ITEMCOUNT = 10;
-   public static final int pageSize = 10;
+   public static final int pageSize = 100;
 
    public static final String ShareID = "5d22bec94ca357c136000062";
    public static final String WX_APPID = "wxed3971fb41174390";
@@ -19,8 +19,10 @@ public class Constants {
    public static final String WB_APPID = "761285940";
    public static final String WB_SECRET = "83c30696e0218437f471581f73b5985c";
 
-   public static final String mainPath = Environment.getExternalStorageDirectory() + "/shegou/";
+   public static final String mainPath = Environment.getExternalStorageDirectory() + "/mugua/";
    public static final String imgUrl = mainPath + "img/";
+   public static final String videoUrl = mainPath + "video/";
+
 
    public static final String ZFB_PAY = "2018102061789029";
 
