@@ -1,16 +1,14 @@
 package com.yc.mugua.view;
 
 import android.os.Bundle;
+import android.support.v7.widget.AppCompatTextView;
 import android.view.View;
 
-import com.lcodecore.tkrefreshlayout.RefreshListenerAdapter;
-import com.lcodecore.tkrefreshlayout.TwinklingRefreshLayout;
 import com.yc.mugua.R;
 import com.yc.mugua.adapter.CashAdapter;
-import com.yc.mugua.adapter.CollectionAdapter;
 import com.yc.mugua.base.BaseFragment;
 import com.yc.mugua.bean.DataBean;
-import com.yc.mugua.databinding.BRecyclerBinding;
+import com.yc.mugua.databinding.FCashBinding;
 import com.yc.mugua.impl.CashContract;
 import com.yc.mugua.presenter.CashPresenter;
 
@@ -24,10 +22,11 @@ import java.util.List;
  * Time: 20:17
  *  缓存
  */
-public class CashFrg extends BaseFragment<CashPresenter, BRecyclerBinding> implements CashContract.View {
+public class CashFrg extends BaseFragment<CashPresenter, FCashBinding> implements CashContract.View {
 
     private List<DataBean> listBean = new ArrayList<>();
     private CashAdapter adapter;
+    private AppCompatTextView topRight;
 
     @Override
     public void initPresenter() {
@@ -41,12 +40,13 @@ public class CashFrg extends BaseFragment<CashPresenter, BRecyclerBinding> imple
 
     @Override
     protected int bindLayout() {
-        return R.layout.b_recycler;
+        return R.layout.f_cash;
     }
 
     @Override
     protected void initView(View view) {
-        setTitle(getString(R.string.local_cache));
+        setTitle(getString(R.string.local_cache), getString(R.string.manage));
+        topRight = view.findViewById(R.id.top_right);
         if (adapter == null){
             adapter = new CashAdapter(act, listBean);
         }
@@ -54,14 +54,16 @@ public class CashFrg extends BaseFragment<CashPresenter, BRecyclerBinding> imple
         mB.recyclerView.setAdapter(adapter);
 
         showLoadDataing();
-        mB.refreshLayout.startRefresh();
+        mB.loadinglayout.setEmptyText("您还没有缓存" + "\n" + "在详情页可点击“离线缓存”按钮进行缓存");
+        mPresenter.onRequest();
+        /*mB.refreshLayout.startRefresh();
         mB.refreshLayout.setEnableLoadmore(false);
         setRefreshLayout(mB.refreshLayout, new RefreshListenerAdapter() {
             @Override
             public void onRefresh(TwinklingRefreshLayout refreshLayout) {
                 mPresenter.onRequest();
             }
-        });
+        });*/
     }
 
     @Override
@@ -84,4 +86,18 @@ public class CashFrg extends BaseFragment<CashPresenter, BRecyclerBinding> imple
         adapter.notifyDataSetChanged();
     }
 
+    @Override
+    protected void setOnRightClickListener() {
+        super.setOnRightClickListener();
+        if (topRight.getText().toString().equals(getString(R.string.manage))){
+            topRight.setText(getString(R.string.cancel));
+            mB.layout.setVisibility(View.VISIBLE);
+            adapter.setVisibbility(true);
+        }else {
+            topRight.setText(getString(R.string.manage));
+            mB.layout.setVisibility(View.GONE);
+            adapter.setVisibbility(false);
+        }
+        adapter.notifyDataSetChanged();
+    }
 }
